algorithms/requirements.txt
pillow
numpy
matplotlib
imageio